+++ template-parts/loop-page.php
@@ -89,7 +89,9 @@ if ( have_posts() ) : while ( have_posts() ) : the_post();
 
   echo '</main>';
 
-  echo lawyerist_cta();
+  if ( !is_really_a_woocommerce_page() ) {
+    echo lawyerist_cta();
+  }
 
   // Shows review template if comments are open and reviews are enabled. The only
   // reason this is present on plain pages is that we're using the regular page