+++ CHANGELOG.md
@@ -7,12 +7,22 @@ All notable changes to this project will be documented in this file. The format 
 
 
 ## [Unreleased]
+
+
+## [5.3.0] - 2020-03-15
+
+### [Added]
+- Added the Small Firm Financial Scorecard to the Small Firm Dashboard.
+- Add LabCon Materials template and praise cards.
+
+### [Changed]
+- Un-echo HTML in template parts.
 - Remove partner dashboard page from signup wall.
 - Don't show call to action on WooCommerce pages.
-- Un-echo HTML in template parts.
+
+### [Fixed]
 - Fix sponsored posts not displaying.
 - Fix bug that was displaying the sticky post container when there weren't any.
-- Add LabCon Materials template and praise cards.
 
 
 ## [5.2.1] - 2020-02-14